Refactor main page into sections
