+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type React from "react"
 import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 import "./globals.css"
-import { TranslationProvider } from "./hooks/use-translation"
+import { Providers } from "./providers"
 
 const inter = Inter({ subsets: ["latin"] })
 
@@ -19,7 +19,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <TranslationProvider>{children}</TranslationProvider>
+        <Providers>{children}</Providers>
       </body>
     </html>
   )
