+++ app/providers.tsx
@@ -1,0 +1,12 @@
+"use client"
+
+import { ThemeProvider } from "next-themes"
+import { TranslationProvider } from "./hooks/use-translation"
+
+export function Providers({ children }: { children: React.ReactNode }) {
+  return (
+    <ThemeProvider attribute="class" defaultTheme="dark">
+      <TranslationProvider>{children}</TranslationProvider>
+    </ThemeProvider>
+  )
+}
