+++ lib/hooks/use-scroll-to-section.tsx
@@ -1,0 +1,8 @@
+"use client"
+
+export function useScrollToSection() {
+  const scrollToSection = (sectionId: string) => {
+    document.getElementById(sectionId)?.scrollIntoView({ behavior: "smooth" })
+  }
+  return { scrollToSection }
+}
