+++ locales/en.json
@@ -1,0 +1,29 @@
+{
+  "navbar.home": "home",
+  "navbar.about": "about_me",
+  "navbar.history": "history",
+  "navbar.projects": "projects",
+  "navbar.contact": "contact",
+  "hero.tags": "Backend | DevOps | Security",
+  "hero.tagline": "Building robust systems one commit at a time",
+  "hero.downloadCV": "Download CV",
+  "about.description": "Backend developer and senior Information Systems Engineering student at UTN FRC. I have hands-on experience building microservices with Java and Spring Boot, QA automation, Docker-based deployments, and modern DevOps practices in cloud environments (AWS, Terraform). I contribute as a teaching assistant in Operating Systems and specialize in integrating security (DevSecOps, Keycloak, Spring Security) and automation within CI/CD pipelines. Currently, I actively participate in cybersecurity challenges and explore the use of AI to automate processes and enhance real-world user experience.",
+  "history.description": "I am currently available and actively seeking new professional challenges where I can apply my skills, continue growing, and contribute real value through the development of robust and secure software.",
+  "projects.apiGateway.description": "High-performance API gateway with rate limiting, authentication, and load balancing",
+  "projects.infrastructure.description": "Terraform modules for AWS infrastructure deployment and management",
+  "projects.security.description": "Automated vulnerability scanner for web applications and APIs",
+  "contact.thanks": "Thanks for visiting!",
+  "timeline.utn.title": "UTN FRC – Operating Systems Teaching Assistant",
+  "timeline.utn.description": "Teaching Assistant in Operating Systems (2024–Present)",
+  "timeline.utn.summary": "Prepared and delivered theoretical and lab classes, graded exams, and provided technical and conceptual support to students. Full course responsibility since 2025.",
+  "timeline.qatitle": "TestingDeSoftwareArg – QA Automation",
+  "timeline.qadescription": "QA Automation, TestingDeSoftwareArg (Jan 2024–Nov 2024)",
+  "timeline.qasummary": "Designed and executed automated tests using Selenium (Python/Java), Locust, TDD, and REST API validation. Integrated testing in CI pipelines.",
+  "timeline.logistics.title": "Logistics Project – Backend Developer (Freelance)",
+  "timeline.logistics.description": "Backend Developer, Freelance Project (Nov 2023–Jan 2024)",
+  "timeline.logistics.summary": "Developed RESTful APIs with Java/Spring Boot, implemented microservice architectures, Docker integration, API Gateway management, and security with Keycloak/Spring Security.",
+  "timeline.soon.year": "2025",
+  "timeline.soon.title": "Coming soon…",
+  "timeline.soon.description": "Preparing for new professional challenges.",
+  "timeline.soon.summary": "Stay tuned for upcoming projects and opportunities."
+}
