+++ locales/es.json
@@ -1,0 +1,29 @@
+{
+  "navbar.home": "inicio",
+  "navbar.about": "sobre_mi",
+  "navbar.history": "historia",
+  "navbar.projects": "proyectos",
+  "navbar.contact": "contacto",
+  "hero.tags": "Backend | DevOps | Seguridad",
+  "hero.tagline": "Construyendo sistemas robustos un commit a la vez",
+  "hero.downloadCV": "Descargar CV",
+  "history.description": "Actualmente me encuentro disponible y en búsqueda activa de nuevos desafíos profesionales donde pueda aplicar mis conocimientos, seguir creciendo y aportar valor real a través del desarrollo de software robusto y seguro.",
+  "about.description": "Desarrollador backend y estudiante avanzado de Ingeniería en Sistemas de Información en UTN FRC. Tengo experiencia real en el desarrollo de microservicios con Java y Spring Boot, automatización de pruebas (QA Automation), despliegue con Docker y prácticas DevOps modernas en entornos cloud (AWS, Terraform). Colaboro como ayudante de cátedra en Sistemas Operativos y me especializo en la integración de seguridad (DevSecOps, Keycloak, Spring Security) y automatización en pipelines CI/CD. Actualmente, participo activamente en desafíos de ciberseguridad y exploro la aplicación de inteligencia artificial para automatizar procesos y mejorar la experiencia de usuario en proyectos reales.",
+  "projects.apiGateway.description": "Gateway de API de alto rendimiento con limitación de velocidad, autenticación y balanceador de carga",
+  "projects.infrastructure.description": "Módulos de Terraform para despliegue y gestión de infraestructura AWS",
+  "projects.security.description": "Escáner automatizado de vulnerabilidades para aplicaciones web y APIs",
+  "contact.thanks": "¡Gracias por visitar!",
+  "timeline.utn.title": "UTN FRC – Ayudante de Sistemas Operativos",
+  "timeline.utn.description": "Ayudante de cátedra en Sistemas Operativos (2024–Actualidad)",
+  "timeline.utn.summary": "Preparé y dicté clases teóricas y prácticas, corregí exámenes y brindé soporte técnico y conceptual a estudiantes. Responsable de un curso completo desde 2025.",
+  "timeline.qatitle": "TestingDeSoftwareArg – QA Automation",
+  "timeline.qadescription": "QA Automation, TestingDeSoftwareArg (Ene 2024–Nov 2024)",
+  "timeline.qasummary": "Diseñé y ejecuté pruebas automatizadas con Selenium (Python/Java), Locust, TDD y validación de APIs REST. Integración de pruebas en pipelines CI.",
+  "timeline.logistics.title": "Proyecto Logística – Backend Freelance",
+  "timeline.logistics.description": "Desarrollador Backend Freelance (Nov 2023–Ene 2024)",
+  "timeline.logistics.summary": "Desarrollé APIs RESTful en Java/Spring Boot, arquitecturas con microservicios, integración con Docker y API Gateway, y seguridad con Keycloak/Spring Security.",
+  "timeline.soon.year": "2025",
+  "timeline.soon.title": "Próximamente…",
+  "timeline.soon.description": "Preparando nuevos desafíos profesionales.",
+  "timeline.soon.summary": "Mantente atento para próximos proyectos y oportunidades."
+}
